storage_courses/pyproject.toml
[tool.poetry]
name = "storage-courses"
version = "0.1.0"
description = ""
authors = ["sky-sky-code <saidyllapashaev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = "^0.20.0"
aerich = "^0.7.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
aiohttp = "^3.9.3"
aio-pika = "^9.4.0"
redis = ">=4.2.0rc1"



[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
